Refactor Demonstration Recorder
