+++ UnitySDK/Assets/ML-Agents/Scripts/DemonstrationRecorder.cs
@@ -4,16 +4,6 @@ using System.IO;
 namespace MLAgents
 {
     /// <summary>
-    /// Demonstration meta-data.
-    /// </summary>
-    [System.Serializable]
-    public struct DemonstrationMetaData
-    {
-        public int numberExperiences;
-        public int numberEpisodes;
-    }
-
-    /// <summary>
     /// Demonstration Recorder Component.
     /// </summary>
     [RequireComponent(typeof(Agent))]
@@ -23,7 +13,7 @@ namespace MLAgents
         public string demonstrationName;
         private Agent recordingAgent;
         private string filePath;
-        private DemonstrationMetaData metaData;
+        private DemonstrationStore demoStore;
 
         /// <summary>
         /// Creates demonstraion file if in Editor mode, and set to record.
@@ -33,61 +23,20 @@ namespace MLAgents
             if (Application.isEditor && record)
             {
                 recordingAgent = GetComponent<Agent>();
-                if (!Directory.Exists("Assets/Demonstrations"))
-                {
-                    Directory.CreateDirectory("Assets/Demonstrations");
-                }
+                demoStore = new DemonstrationStore();
+                demoStore.CreateDirectory();                
 
-                metaData = new DemonstrationMetaData
-                {
-                    numberEpisodes = 0,
-                    numberExperiences = 0
-                };
-                CreateDemonstrationFile();
+                demoStore.CreateDemonstrationFile(demonstrationName);
+                demoStore.WriteBrainParameters(recordingAgent.brain.brainParameters);
             }
         }
 
         /// <summary>
-        /// Creates demonstration file, and writes brainParameters as json to file.
+        /// Forwards AgentInfo to Demonstration Store.
         /// </summary>
-        private void CreateDemonstrationFile()
-        {
-            // Creates demonstration file.
-            var literalName = demonstrationName;
-            filePath = "Assets/Demonstrations/" + literalName + ".demo";
-            var uniqueNameCounter = 0;
-            while (File.Exists(filePath))
-            {
-                literalName = demonstrationName + "_" + uniqueNameCounter;
-                filePath = "Assets/Demonstrations/" + literalName + ".demo";
-                uniqueNameCounter++;
-            }
-
-            // Writes BrainParameters to file.
-            var jsonParameters = JsonUtility.ToJson(recordingAgent.brain.brainParameters);
-            var writer = File.CreateText(filePath);
-            writer.Write(jsonParameters + '\n');
-            writer.Close();
-        }
-
-        /// <summary>
-        /// Write AgentInfo experience to file as json.
-        /// </summary>
-        /// <param name="info">AgentInfo of current experience</param>
         public void WriteExperience(AgentInfo info)
         {
-            // Increment meta-data counters.
-            metaData.numberExperiences++;
-            if (info.done)
-            {
-                metaData.numberEpisodes++;
-            }
-
-            // Write AgentInfo to file.
-            var jsonInfo = JsonUtility.ToJson(info);
-            var writer = new StreamWriter(filePath, true);
-            writer.WriteLine(jsonInfo);
-            writer.Close();
+            demoStore.WriteExperience(info);
         }
 
         /// <summary>
@@ -97,11 +46,7 @@ namespace MLAgents
         {
             if (Application.isEditor && record)
             {
-                // Write meta-data to file.
-                var jsonInfo = JsonUtility.ToJson(metaData);
-                var writer = new StreamWriter(filePath, true);
-                writer.WriteLine(jsonInfo);
-                writer.Close();
+                demoStore.WriteMetadata();
             }
         }
     }
